Move c_str! macro to utils + move Janus callbacks to separate mod
