+++ src/utils.rs
@@ -1,0 +1,7 @@
+// courtesy of c_string crate, which also has some other stuff we aren't interested in
+// taking in as a dependency here.
+macro_rules! c_str {
+    ($lit:expr) => {
+        unsafe { CStr::from_ptr(concat!($lit, "\0").as_ptr() as *const $crate::c_char) }
+    };
+}
